Refactor balanced strategy into modules
