+++ mexc_bot/core/balanced_strategy_base.py
@@ -10,6 +10,18 @@ import logging
 from typing import Optional, Dict, Any
 from dataclasses import dataclass
 from telegram_utils import tg_send
+from .indicators import (
+    calculate_ema,
+    calculate_rsi,
+    calculate_atr,
+    calculate_adx,
+    calculate_macd,
+)
+from .backtest import run_backtest as backtest_run
+from .plots import (
+    plot_equity_curve as plot_equity_curve_func,
+    plot_regime_performance as plot_regime_performance_func,
+)
 
 # Магические числа
 COMMISSION_RATE_ENTRY = 0.00035  # 0.035% комиссия при входе
@@ -787,71 +799,19 @@ class BalancedAdaptiveStrategy:
         }
     
     def _calculate_ema(self, series, period):
-        return series.ewm(span=period, adjust=False).mean()
-    
+        return calculate_ema(series, period)
+
     def _calculate_rsi(self, series, period):
-        delta = series.diff()
-        gain = delta.where(delta > 0, 0)
-        loss = -delta.where(delta < 0, 0)
-        avg_gain = gain.rolling(window=period).mean()
-        avg_loss = loss.rolling(window=period).mean()
-        rs = avg_gain / avg_loss.clip(lower=1e-10)  # (6) Исправление replace на clip
-        rsi = 100 - (100 / (1 + rs))
-        return rsi
-    
+        return calculate_rsi(series, period)
+
     def _calculate_atr(self, high, low, close, period):
-        tr1 = high - low
-        tr2 = (high - close.shift(1)).abs()
-        tr3 = (low - close.shift(1)).abs()
-        tr = pd.DataFrame({'tr1': tr1, 'tr2': tr2, 'tr3': tr3}).max(axis=1)
-        
-        atr = tr.rolling(window=period).mean()
-        return atr
-    
+        return calculate_atr(high, low, close, period)
+
     def _calculate_adx(self, high, low, close, period):
-        tr1 = high - low
-        tr2 = (high - close.shift(1)).abs()
-        tr3 = (low - close.shift(1)).abs()
-        tr = pd.DataFrame({'tr1': tr1, 'tr2': tr2, 'tr3': tr3}).max(axis=1)
-        atr = tr.rolling(window=period).mean()
-        
-        up_move = high - high.shift(1)
-        down_move = low.shift(1) - low
-        
-        plus_dm = np.where((up_move > down_move) & (up_move > 0), up_move, 0)
-        minus_dm = np.where((down_move > up_move) & (down_move > 0), down_move, 0)
-        
-        plus_dm = pd.Series(plus_dm, index=high.index)
-        minus_dm = pd.Series(minus_dm, index=high.index)
-        
-        smooth_plus_dm = plus_dm.rolling(window=period).mean()
-        smooth_minus_dm = minus_dm.rolling(window=period).mean()
-        
-        plus_di = 100 * (smooth_plus_dm / atr.replace(0, 1e-10))
-        minus_di = 100 * (smooth_minus_dm / atr.replace(0, 1e-10))
-        
-        dx = 100 * ((plus_di - minus_di).abs() / (plus_di + minus_di).replace(0, 1e-10))
-        adx = dx.rolling(window=period).mean()
-        
-        return {
-            'ADX': adx,
-            'Plus_DI': plus_di,
-            'Minus_DI': minus_di
-        }
-    
+        return calculate_adx(high, low, close, period)
+
     def _calculate_macd(self, series, fast_period, slow_period, signal_period):
-        fast_ema = series.ewm(span=fast_period, adjust=False).mean()
-        slow_ema = series.ewm(span=slow_period, adjust=False).mean()
-        
-        macd = fast_ema - slow_ema
-        macd_signal = macd.ewm(span=signal_period, adjust=False).mean()
-        macd_hist = macd - macd_signal
-        
-        return {
-            'MACD': macd,
-            'MACD_Signal': macd_signal,
-            'MACD_Hist': macd_hist
-        }
+        return calculate_macd(series, fast_period, slow_period, signal_period)
     
     def get_trading_signals(self, current, previous, regime_type):
         long_signals = []
@@ -1490,592 +1450,7 @@ class BalancedAdaptiveStrategy:
 
     def run_backtest(self) -> Optional[pd.DataFrame]:
         """Run the backtest and return results DataFrame"""
-        logging.info("Running backtest...")
-        balance = self.initial_balance
-        position = 0
-        entry_price = 0.0
-        position_size = 0.0
-        entry_date = None
-        last_trade_date = None
-        pyramid_entries = 0
-        stop_loss_price = 0.0
-        take_profit_price = 0.0
-        max_trade_price = 0.0
-        min_trade_price = float('inf')
-        results = []
-        self.trade_history = []
-        self.recent_long_win_rate = 0.5
-        self.recent_short_win_rate = 0.5
-        short_ema = self.params['short_ema']
-        long_ema = self.params['long_ema']
-        for i in range(1, len(self.data)):
-            current = self.data.iloc[i]
-            previous = self.data.iloc[i-1]
-            current_balance = balance
-            current_equity = balance
-            unrealized_pnl_pct = 0.0
-            if position != 0:
-                if entry_price <= 0:
-                    position = 0
-                    entry_price = 0.0
-                    position_size = 0.0
-                    entry_date = None
-                    pyramid_entries = 0
-                    max_trade_price = 0.0
-                    min_trade_price = float('inf')
-                    continue
-                max_trade_price = max(max_trade_price, current['High'])
-                min_trade_price = min(min_trade_price, current['Low'])
-                trade_age_hours = (current.name - entry_date).total_seconds() / 3600
-                higher_tf_bullish = current['Higher_TF_Bullish']
-                if position == 1:
-                    optimal_leverage = self.calculate_optimal_leverage(current, 'LONG', self.max_leverage)
-                    unrealized_pnl_pct = (current['Close'] / entry_price) - 1
-                    unrealized_pnl = unrealized_pnl_pct * position_size
-                    current_equity = balance + unrealized_pnl
-                    
-                    new_stop = self.apply_trailing_stop('LONG', entry_price, current['Close'], max_trade_price, min_trade_price, unrealized_pnl_pct)
-                    if new_stop is not None and new_stop > stop_loss_price:
-                        stop_loss_price = new_stop
-                    
-                    if current['Low'] <= stop_loss_price:
-                        pnl = self._close_position('LONG', position_size, entry_price, stop_loss_price)
-                        balance += pnl
-                        
-                        self.trade_history[-1].update({
-                            'exit_date': current.name,
-                            'exit_price': stop_loss_price,
-                            'pnl': pnl,
-                            'balance': balance,
-                            'reason': 'Stop Loss',
-                            'pyramid_entries': pyramid_entries,
-                            'trade_duration': trade_age_hours
-                        })
-                        self._notify_trade_close(self.trade_history[-1])
-                        
-                        position = 0
-                        entry_price = 0.0
-                        position_size = 0.0
-                        entry_date = None
-                        last_trade_date = current.name
-                        pyramid_entries = 0
-                        max_trade_price = 0.0
-                        min_trade_price = float('inf')
-                        continue
-                    
-                    if current['High'] >= take_profit_price:
-                        pnl = self._close_position('LONG', position_size, entry_price, take_profit_price)
-                        balance += pnl
-                        
-                        self.trade_history[-1].update({
-                            'exit_date': current.name,
-                            'exit_price': take_profit_price,
-                            'pnl': pnl,
-                            'balance': balance,
-                            'reason': 'Take Profit',
-                            'pyramid_entries': pyramid_entries,
-                            'trade_duration': trade_age_hours
-                        })
-                        self._notify_trade_close(self.trade_history[-1])
-                        
-                        position = 0
-                        entry_price = 0.0
-                        position_size = 0.0
-                        entry_date = None
-                        last_trade_date = current.name
-                        pyramid_entries = 0
-                        max_trade_price = 0.0
-                        min_trade_price = float('inf')
-                        continue
-                    
-                    if (pyramid_entries < self.params['max_pyramid_entries'] and 
-                        current['Bullish_Trend'] and 
-                        current['ADX'] > 40 and 
-                        current['Close'] > entry_price * (1 + self.params['pyramid_min_profit']) and
-                        (current['MACD_Bullish_Cross'] or current['Final_Long_Bias'] > 0.7)):
-                        
-                        additional_size = position_size * self.params['pyramid_size_multiplier']
-                        old_value = entry_price * position_size
-                        new_value = current['Close'] * additional_size
-                        position_size += additional_size
-                        entry_price = (old_value + new_value) / position_size
-                        
-                        exit_levels = self.calculate_dynamic_exit_levels('LONG', entry_price, current, trade_age_hours)
-                        stop_loss_price = exit_levels['stop_loss']
-                        take_profit_price = exit_levels['take_profit']
-                        
-                        pyramid_entries += 1
-                        continue
-                    
-                    exit_signals = []
-                    current_regime = 'trend' if current['Trend_Weight'] > 0.5 else 'range'
-                    
-                    if current_regime == 'trend':
-                        if ((previous[f'EMA_{short_ema}'] >= previous[f'EMA_{long_ema}']) and 
-                            (current[f'EMA_{short_ema}'] < current[f'EMA_{long_ema}'])):
-                            exit_signals.append('EMA Crossover')
-                        
-                        if current['MACD_Bearish_Cross']:
-                            exit_signals.append('MACD Crossover')
-                        
-                        if (current['Bearish_Trend'] and not previous['Bearish_Trend']):
-                            exit_signals.append('Trend Change')
-                        
-                        if current['Higher_TF_Bearish'] and unrealized_pnl_pct > 0.02:
-                            exit_signals.append('Higher TF Trend Change')
-                        
-                        if 'Market_Regime' in current:
-                            if current['Market_Regime'] == 'transition_to_bear' and unrealized_pnl_pct > 0.02:
-                                exit_signals.append('Regime Change to Bear')
-                    
-                    else:
-                        if current['RSI'] > self.params['rsi_overbought']:
-                            exit_signals.append('RSI Overbought')
-                        if current['Close'] > current['BB_Upper']:
-                            exit_signals.append('Upper Bollinger')
-                        if current['Bearish_Divergence']:
-                            exit_signals.append('Bearish Divergence')
-                        if current['Bearish_Engulfing'] and unrealized_pnl_pct > 0.02:
-                            exit_signals.append('Bearish Engulfing')
-                        if current['Stat_Overbought'] and unrealized_pnl_pct > 0.03:
-                            exit_signals.append('Statistical Overbought')
-                    
-                    if current['Volatility_Ratio'] > 2.0 and unrealized_pnl_pct > 0.03:
-                        exit_signals.append('Volatility Spike')
-                    
-                    if (unrealized_pnl_pct > 0.08 and 
-                        previous['Close'] > current['Close'] and 
-                        previous['MACD'] > current['MACD']):
-                        exit_signals.append('Profit Protection')
-                    
-                    if 'Potential_Momentum_Reversal' in current and current['Potential_Momentum_Reversal'] and unrealized_pnl_pct > 0.04:
-                        exit_signals.append('Momentum Reversal')
-                    
-                    if 'Final_Short_Bias' in current and current['Final_Short_Bias'] > 0.7 and unrealized_pnl_pct > 0.03:
-                        exit_signals.append('Bias Shift')
-                    
-                    if exit_signals and trade_age_hours > 4:
-                        pnl = self._close_position('LONG', position_size, entry_price, current['Close'])
-                        balance += pnl
-                        
-                        self.trade_history[-1].update({
-                            'exit_date': current.name,
-                            'exit_price': current['Close'],
-                            'pnl': pnl,
-                            'balance': balance,
-                            'reason': ', '.join(exit_signals),
-                            'pyramid_entries': pyramid_entries,
-                            'trade_duration': trade_age_hours
-                        })
-                        self._notify_trade_close(self.trade_history[-1])
-                        
-                        position = 0
-                        entry_price = 0.0
-                        position_size = 0.0
-                        entry_date = None
-                        last_trade_date = current.name
-                        pyramid_entries = 0
-                        max_trade_price = 0.0
-                        min_trade_price = float('inf')
-                        continue
-                
-                elif position == -1:
-                    optimal_leverage = self.calculate_optimal_leverage(current, 'SHORT', self.max_leverage)
-                    unrealized_pnl_pct = 1 - (current['Close'] / entry_price)
-                    unrealized_pnl = unrealized_pnl_pct * position_size
-                    current_equity = balance + unrealized_pnl
-                    
-                    new_stop = self.apply_trailing_stop('SHORT', entry_price, current['Close'], max_trade_price, min_trade_price, unrealized_pnl_pct)
-                    if new_stop is not None and new_stop < stop_loss_price:
-                        stop_loss_price = new_stop
-                    
-                    if current['High'] >= stop_loss_price:
-                        pnl = self._close_position('SHORT', position_size, entry_price, stop_loss_price)
-                        balance += pnl
-                        
-                        self.trade_history[-1].update({
-                            'exit_date': current.name,
-                            'exit_price': stop_loss_price,
-                            'pnl': pnl,
-                            'balance': balance,
-                            'reason': 'Stop Loss',
-                            'pyramid_entries': pyramid_entries,
-                            'trade_duration': trade_age_hours
-                        })
-                        self._notify_trade_close(self.trade_history[-1])
-
-                        position = 0
-                        entry_price = 0.0
-                        position_size = 0.0
-                        entry_date = None
-                        last_trade_date = current.name
-                        pyramid_entries = 0
-                        max_trade_price = 0.0
-                        min_trade_price = float('inf')
-                        continue
-                    
-                    if current['Low'] <= take_profit_price:
-                        pnl = self._close_position('SHORT', position_size, entry_price, take_profit_price)
-                        balance += pnl
-                        
-                        self.trade_history[-1].update({
-                            'exit_date': current.name,
-                            'exit_price': take_profit_price,
-                            'pnl': pnl,
-                            'balance': balance,
-                            'reason': 'Take Profit',
-                            'pyramid_entries': pyramid_entries,
-                            'trade_duration': trade_age_hours
-                        })
-                        self._notify_trade_close(self.trade_history[-1])
-                        position = 0
-                        entry_price = 0.0
-                        position_size = 0.0
-                        entry_date = None
-                        last_trade_date = current.name
-                        pyramid_entries = 0
-                        max_trade_price = 0.0
-                        min_trade_price = float('inf')
-                        continue
-                    
-                    if (pyramid_entries < self.params['max_pyramid_entries'] and 
-                        current['Bearish_Trend'] and 
-                        current['ADX'] > 40 and 
-                        current['Close'] < entry_price * (1 - self.params['pyramid_min_profit']) and
-                        (current['MACD_Bearish_Cross'] or current['Final_Short_Bias'] > 0.7)):
-                        
-                        additional_size = position_size * self.params['pyramid_size_multiplier']
-                        old_value = entry_price * position_size
-                        new_value = current['Close'] * additional_size
-                        position_size += additional_size
-                        entry_price = (old_value + new_value) / position_size
-                        
-                        exit_levels = self.calculate_dynamic_exit_levels('SHORT', entry_price, current, trade_age_hours)
-                        stop_loss_price = exit_levels['stop_loss']
-                        take_profit_price = exit_levels['take_profit']
-                        
-                        pyramid_entries += 1
-                        continue
-                    
-                    exit_signals = []
-                    current_regime = 'trend' if current['Trend_Weight'] > 0.5 else 'range'
-                    
-                    if current_regime == 'trend':
-                        if ((previous[f'EMA_{short_ema}'] <= previous[f'EMA_{long_ema}']) and 
-                            (current[f'EMA_{short_ema}'] > current[f'EMA_{long_ema}'])):
-                            exit_signals.append('EMA Crossover')
-                        
-                        if current['MACD_Bullish_Cross']:
-                            exit_signals.append('MACD Crossover')
-                        
-                        if current['Bullish_Trend'] and not previous['Bullish_Trend']:
-                            exit_signals.append('Trend Change')
-                        
-                        if current['Higher_TF_Bullish'] and unrealized_pnl_pct > 0.02:
-                            exit_signals.append('Higher TF Trend Change')
-                        
-                        if 'Market_Regime' in current:
-                            if current['Market_Regime'] == 'transition_to_bull' and unrealized_pnl_pct > 0.02:
-                                exit_signals.append('Regime Change to Bull')
-                    
-                    else:
-                        if current['RSI'] < self.params['rsi_oversold']:
-                            exit_signals.append('RSI Oversold')
-                        if current['Close'] < current['BB_Lower']:
-                            exit_signals.append('Lower Bollinger')
-                        if current['Bullish_Divergence']:
-                            exit_signals.append('Bullish Divergence')
-                        if current['Bullish_Engulfing'] and unrealized_pnl_pct > 0.02:
-                            exit_signals.append('Bullish Engulfing')
-                        if current['Stat_Oversold'] and unrealized_pnl_pct > 0.03:
-                            exit_signals.append('Statistical Oversold')
-                    
-                    if current['Volatility_Ratio'] > 2.0 and unrealized_pnl_pct > 0.03:
-                        exit_signals.append('Volatility Spike')
-                    
-                    if (unrealized_pnl_pct > 0.08 and 
-                        previous['Close'] < current['Close'] and 
-                        previous['MACD'] < current['MACD']):
-                        exit_signals.append('Profit Protection')
-                    
-                    if 'Potential_Momentum_Reversal' in current and current['Potential_Momentum_Reversal'] and unrealized_pnl_pct > 0.04:
-                        exit_signals.append('Momentum Reversal')
-                    
-                    if 'Final_Long_Bias' in current and current['Final_Long_Bias'] > 0.7 and unrealized_pnl_pct > 0.03:
-                        exit_signals.append('Bias Shift')
-                    
-                    if exit_signals and trade_age_hours > 4:
-                        pnl = self._close_position('SHORT', position_size, entry_price, current['Close'])
-                        balance += pnl
-                        
-                        self.trade_history[-1].update({
-                            'exit_date': current.name,
-                            'exit_price': current['Close'],
-                            'pnl': pnl,
-                            'balance': balance,
-                            'reason': ', '.join(exit_signals),
-                            'pyramid_entries': pyramid_entries,
-                            'trade_duration': trade_age_hours
-                        })
-                        self._notify_trade_close(self.trade_history[-1])
-
-                        position = 0
-                        entry_price = 0.0
-                        position_size = 0.0
-                        entry_date = None
-                        last_trade_date = current.name
-                        pyramid_entries = 0
-                        max_trade_price = 0.0
-                        min_trade_price = float('inf')
-                        continue
-            
-            if position == 0:
-                if last_trade_date is not None:
-                    hours_since_last_trade = (current.name - last_trade_date).total_seconds() / 3600
-                    if hours_since_last_trade < self.min_trades_interval:
-                        continue
-                
-                if not current['Active_Hours']:
-                    continue
-                
-                if not self.is_optimal_trading_time(current.name):
-                    continue
-                
-                current_regime = 'trend' if current['Trend_Weight'] > 0.5 else 'range'
-                signals = self.get_trading_signals(current, previous, current_regime)
-                filtered_signals = self.apply_advanced_filtering(current, signals)
-                
-                market_regime = current.get('Market_Regime', 'unknown')
-                
-                adaptive_risk = self.adaptive_risk_per_trade(
-                    market_regime, 
-                    self.recent_long_win_rate, 
-                    self.recent_short_win_rate
-                )
-                
-                min_signal_threshold = 0.7  # Было 0.6, стало 0.7
-                
-                # --- ENTRY LOGIC ---
-                long_entry_threshold = 0.60
-                short_entry_threshold = 0.60
-                # --- Volatility adjustment ---
-                volatility_adjustment = current['ATR'] / current['ATR_MA'] if current['ATR_MA'] else 1
-                if volatility_adjustment > 1.2:
-                    long_entry_threshold *= 0.9
-                    short_entry_threshold *= 0.9
-                # --- Hour-based adjustment ---
-                optimal_hours = [9, 10, 13, 14, 15, 16]
-                if current.name.hour in optimal_hours:
-                    signal_threshold_multiplier = 0.9
-                else:
-                    signal_threshold_multiplier = 1.2
-                long_entry_threshold *= signal_threshold_multiplier
-                short_entry_threshold *= signal_threshold_multiplier
-                #
-                if (filtered_signals['long_weight'] > filtered_signals['short_weight'] and 
-                    filtered_signals['long_weight'] >= long_entry_threshold):
-                    self.current_side = 'LONG'
-                    if not self.is_optimal_trading_time(current.name):
-                        continue
-                    position = 1
-                    entry_price = current['Close']
-                    if entry_price <= 0:
-                        entry_price = 1.0
-                    entry_date = current.name
-                    
-                    optimal_leverage = self.calculate_optimal_leverage(current, 'LONG', self.max_leverage)
-                    exit_levels = self.calculate_dynamic_exit_levels('LONG', entry_price, current)
-                    stop_loss_price = exit_levels['stop_loss']
-                    take_profit_price = exit_levels['take_profit']
-                    
-                    risk_per_trade = adaptive_risk['LONG']
-                    position_size = self.adaptive_position_sizing(balance, risk_per_trade, entry_price, stop_loss_price, optimal_leverage)
-                    entry_commission = position_size * COMMISSION_RATE_ENTRY
-                    balance -= entry_commission
-
-                    max_trade_price = current['High']
-                    min_trade_price = current['Low']
-                    
-                    signal_info = ', '.join(signal for signal, _ in signals['long_signals'])
-                    
-                    self.trade_history.append({
-                        'entry_date': entry_date,
-                        'entry_price': entry_price,
-                        'exit_date': None,
-                        'exit_price': None,
-                        'position': 'LONG',
-                        'pnl': None,
-                        'balance': balance,
-                        'reason': f'Entry: {signal_info}',
-                        'stop_loss': stop_loss_price,
-                        'take_profit': take_profit_price,
-                        'weight': filtered_signals['long_weight'],
-                        'risk_per_trade': risk_per_trade,
-                        'leverage': optimal_leverage,
-                        'market_regime': market_regime,
-                        'market_health': current.get('Market_Health', None),
-                        'position_size': position_size,
-                        'pyramid_entries': pyramid_entries
-                    })
-                    self._notify_trade_open(self.trade_history[-1])
-                
-                elif (filtered_signals['short_weight'] > filtered_signals['long_weight'] and 
-                      filtered_signals['short_weight'] >= short_entry_threshold):
-                    self.current_side = 'SHORT'
-                    if not self.is_optimal_trading_time(current.name):
-                        continue
-                    position = -1
-                    entry_price = current['Close']
-                    if entry_price <= 0:
-                        entry_price = 1.0
-                    entry_date = current.name
-                    
-                    optimal_leverage = self.calculate_optimal_leverage(current, 'SHORT', self.max_leverage)
-                    exit_levels = self.calculate_dynamic_exit_levels('SHORT', entry_price, current)
-                    stop_loss_price = exit_levels['stop_loss']
-                    take_profit_price = exit_levels['take_profit']
-                    
-                    risk_per_trade = adaptive_risk['SHORT']
-                    position_size = self.adaptive_position_sizing(balance, risk_per_trade, entry_price, stop_loss_price, optimal_leverage)
-                    entry_commission = position_size * COMMISSION_RATE_ENTRY
-                    balance -= entry_commission
-
-                    max_trade_price = current['High']
-                    min_trade_price = current['Low']
-                    
-                    signal_info = ', '.join(signal for signal, _ in signals['short_signals'])
-                    
-                    self.trade_history.append({
-                        'entry_date': entry_date,
-                        'entry_price': entry_price,
-                        'exit_date': None,
-                        'exit_price': None,
-                        'position': 'SHORT',
-                        'pnl': None,
-                        'balance': balance,
-                        'reason': f'Entry: {signal_info}',
-                        'stop_loss': stop_loss_price,
-                        'take_profit': take_profit_price,
-                        'weight': filtered_signals['short_weight'],
-                        'risk_per_trade': risk_per_trade,
-                        'leverage': optimal_leverage,
-                        'market_regime': market_regime,
-                        'market_health': current.get('Market_Health', None),
-                        'position_size': position_size,
-                        'pyramid_entries': pyramid_entries
-                    })
-                    self._notify_trade_open(self.trade_history[-1])
-            
-            results.append({
-                'date': current.name,
-                'close': current['Close'],
-                'balance': balance,
-                'equity': current_equity,
-                'position': position,
-                'entry_price': entry_price,
-                'position_size': position_size,
-                'rsi': current['RSI'],
-                'adx': current['ADX'],
-                'is_trending': current['Strong_Trend'],
-                'is_ranging': current['Weak_Trend'],
-                'trend_weight': current['Trend_Weight'],
-                'range_weight': current['Range_Weight'],
-                'market_regime': current.get('Market_Regime', 'unknown'),
-                'market_health': current.get('Market_Health', None),
-                'long_bias': current.get('Final_Long_Bias', None),
-                'short_bias': current.get('Final_Short_Bias', None)
-            })
-            
-            if len(self.trade_history) % 20 == 0 and len(self.trade_history) > 0:
-                self.dynamically_adjust_risk_parameters()
-            
-            if i % 672 == 0 and len(self.trade_history) > 30:
-                self.rebalance_long_short_bias()
-            
-            # --- PARTIAL EXIT LOGIC ---
-            if position == 1 and 'unrealized_pnl_pct' in locals() and unrealized_pnl_pct > 0.12 and position_size > 0:
-                partial_size = position_size * 0.4
-                partial_pnl = partial_size * ((current['Close'] / entry_price) - 1)
-                commission = partial_size * COMMISSION_RATE_EXIT
-                slippage = partial_size * self.slippage_pct / 100
-                partial_pnl -= (commission + slippage)
-                balance += partial_pnl
-                position_size -= partial_size
-            if position == -1 and 'unrealized_pnl_pct' in locals() and unrealized_pnl_pct > 0.12 and position_size > 0:
-                partial_size = position_size * 0.4
-                partial_pnl = partial_size * (1 - (current['Close'] / entry_price))
-                commission = partial_size * COMMISSION_RATE_EXIT
-                slippage = partial_size * self.slippage_pct / 100
-                partial_pnl -= (commission + slippage)
-                balance += partial_pnl
-                position_size -= partial_size
-            
-            # --- Пирамидинг для LONG ---
-            if (position == 1 and
-                pyramid_entries < self.params['max_pyramid_entries'] and 
-                current['Bullish_Trend'] and 
-                current['ADX'] > 30 and 
-                current['Close'] > entry_price * (1 + 0.02) and 
-                (current['MACD_Bullish_Cross'] or current['Final_Long_Bias'] > 0.65)):
-                additional_size = position_size * self.params['pyramid_size_multiplier']
-                old_value = entry_price * position_size
-                new_value = current['Close'] * additional_size
-                position_size += additional_size
-                balance -= additional_size * COMMISSION_RATE_ENTRY
-                entry_price = (old_value + new_value) / position_size
-                exit_levels = self.calculate_dynamic_exit_levels('LONG', entry_price, current)
-                stop_loss_price = exit_levels['stop_loss']
-                take_profit_price = exit_levels['take_profit']
-                pyramid_entries += 1
-                continue
-        
-        # Закрыть позицию в конце
-        if position != 0 and entry_price > 0:
-            last_candle = self.data.iloc[-1]
-            exit_price = last_candle['Close']
-            trade_age_hours = (last_candle.name - entry_date).total_seconds() / 3600
-            
-            if position == 1:
-                pnl = self._close_position('LONG', position_size, entry_price, exit_price)
-            else:
-                pnl = self._close_position('SHORT', position_size, entry_price, exit_price)
-            balance += pnl
-            
-            for trade in reversed(self.trade_history):
-                if trade['exit_date'] is None:
-                    trade['exit_date'] = last_candle.name
-                    trade['exit_price'] = exit_price
-                    trade['pnl'] = pnl
-                    trade['balance'] = balance
-                    trade['reason'] = trade['reason'] + ', End of Backtest'
-                    trade['trade_duration'] = trade_age_hours
-                    break
-        
-        # После полного выхода сбрасываем current_side
-        self.current_side = None
-        self.backtest_results = pd.DataFrame(results)
-        
-        if self.trade_history:
-            self.trade_df = pd.DataFrame(self.trade_history)
-            self.trade_df['exit_date'].fillna(self.data.index[-1], inplace=True)
-            self.trade_df['exit_price'].fillna(self.data['Close'].iloc[-1], inplace=True)
-            mask = self.trade_df['pnl'].isna()
-            self.trade_df.loc[mask & (self.trade_df['position']=='LONG'), 'pnl'] = (
-                self.trade_df.loc[mask & (self.trade_df['position']=='LONG'), 'position_size'] *
-                (self.trade_df.loc[mask & (self.trade_df['position']=='LONG'), 'exit_price'] / self.trade_df.loc[mask & (self.trade_df['position']=='LONG'), 'entry_price'] - 1)
-            )
-            self.trade_df.loc[mask & (self.trade_df['position']=='SHORT'), 'pnl'] = (
-                self.trade_df.loc[mask & (self.trade_df['position']=='SHORT'), 'position_size'] *
-                (1 - self.trade_df.loc[mask & (self.trade_df['position']=='SHORT'), 'exit_price'] / self.trade_df.loc[mask & (self.trade_df['position']=='SHORT'), 'entry_price'])
-            )
-        else:
-            self.trade_df = pd.DataFrame()
-        
-        self.analyze_hour_performance()
-        self.analyze_day_performance()
-        
-        logging.info("Backtest completed")
-        return self.backtest_results
+        return backtest_run(self)
     
     def rebalance_long_short_bias(self):
         if not hasattr(self, 'trade_df') or self.trade_df is None or len(self.trade_df) < 10:
@@ -2089,207 +1464,10 @@ class BalancedAdaptiveStrategy:
             self.params['global_short_penalty'] = 1.0
     
     def plot_equity_curve(self):
-        if self.backtest_results is None:
-            print("No backtest results available. Run backtest first.")
-            return
-        
-        fig, (ax1, ax2, ax3) = plt.subplots(3, 1, figsize=(14, 12), gridspec_kw={'height_ratios': [3, 1, 1]})
-        
-        ax1.plot(self.backtest_results['date'], self.backtest_results['equity'], label='Equity', color='blue')
-        ax1.plot(self.backtest_results['date'], self.backtest_results['balance'], label='Balance', color='green')
-        
-        equity_curve = self.backtest_results['equity']
-        running_max = equity_curve.cummax()
-        drawdown = (running_max - equity_curve) / running_max * 100
-        
-        ax2.fill_between(self.backtest_results['date'], 0, drawdown, color='red', alpha=0.3)
-        ax2.set_ylim(bottom=0, top=max(drawdown) * 1.5)
-        ax2.invert_yaxis()
-        
-        if 'market_health' in self.backtest_results.columns and not self.backtest_results['market_health'].isnull().all():
-            ax3.plot(self.backtest_results['date'], self.backtest_results['market_health'], 
-                     label='Market Health', color='purple', alpha=0.7)
-            ax3.set_ylim(bottom=0, top=100)
-            ax3.axhline(y=50, color='gray', linestyle='--', alpha=0.5)
-            ax3.set_title('Market Health Score (0-100)')
-        else:
-            if 'market_regime' in self.backtest_results.columns and not self.backtest_results['market_regime'].isnull().all():
-                regime_indicator = self.backtest_results['market_regime'].map({
-                    'strong_bull': 1.0,
-                    'transition_to_bull': 0.5,
-                    'choppy_range': 0,
-                    'transition_to_bear': -0.5,
-                    'strong_bear': -1.0,
-                    'mixed': 0,
-                    'unknown': np.nan
-                }).fillna(0)
-                
-                ax3.plot(self.backtest_results['date'], regime_indicator, 
-                         label='Market Regime', color='orange', alpha=0.7)
-                ax3.set_ylim(bottom=-1.2, top=1.2)
-                ax3.axhline(y=0, color='gray', linestyle='--', alpha=0.5)
-                ax3.set_title('Market Regime (1=Bull, 0=Range, -1=Bear)')
-        
-        date_format = mdates.DateFormatter('%Y-%m-%d')
-        ax1.xaxis.set_major_formatter(date_format)
-        ax2.xaxis.set_major_formatter(date_format)
-        ax3.xaxis.set_major_formatter(date_format)
-        
-        def currency_formatter(x, pos):
-            return f'${x:,.0f}'
-        ax1.yaxis.set_major_formatter(FuncFormatter(currency_formatter))
-        
-        def percentage_formatter(x, pos):
-            return f'{x:.0f}%'
-        ax2.yaxis.set_major_formatter(FuncFormatter(percentage_formatter))
-        
-        initial_balance = self.initial_balance
-        final_balance = self.backtest_results['balance'].iloc[-1]
-        total_return = ((final_balance / initial_balance) - 1) * 100
-        max_drawdown_val = drawdown.max()
-        
-        if hasattr(self, 'trade_df') and self.trade_df is not None and len(self.trade_df) > 0:
-            win_rate = len(self.trade_df[self.trade_df['pnl'] > 0]) / len(self.trade_df) * 100
-        else:
-            win_rate = 0
-        
-        textstr = '\n'.join((
-            f'Initial Balance: ${initial_balance:,.2f}',
-            f'Final Balance: ${final_balance:,.2f}',
-            f'Total Return: {total_return:.2f}%',
-            f'Max Drawdown: {max_drawdown_val:.2f}%',
-            f'Win Rate: {win_rate:.2f}%',
-            f'Total Trades: {len(self.trade_df) if hasattr(self, "trade_df") else 0}'
-        ))
-        
-        props = dict(boxstyle='round', facecolor='white', alpha=0.5)
-        ax1.text(0.02, 0.05, textstr, transform=ax1.transAxes, fontsize=10,
-                 verticalalignment='bottom', bbox=props)
-        
-        if hasattr(self, 'trade_df') and len(self.trade_df) > 0:
-            sample_trades = self.trade_df.sample(min(50, len(self.trade_df))) if len(self.trade_df) > 50 else self.trade_df
-            for idx, trade in sample_trades.iterrows():
-                if trade['position'] == 'LONG':
-                    color = 'green'
-                    marker = '^'
-                else:
-                    color = 'red'
-                    marker = 'v'
-                
-                entry_date = pd.to_datetime(trade['entry_date'])
-                exit_date = pd.to_datetime(trade['exit_date'])
-                
-                closest_entry = min(self.backtest_results['date'], key=lambda x: abs(x - entry_date))
-                entry_equity = self.backtest_results.loc[self.backtest_results['date'] == closest_entry, 'equity'].values
-                if len(entry_equity) > 0:
-                    ax1.scatter(closest_entry, entry_equity[0], color=color, s=50, marker=marker, alpha=0.6)
-                
-                closest_exit = min(self.backtest_results['date'], key=lambda x: abs(x - exit_date))
-                exit_equity = self.backtest_results.loc[self.backtest_results['date'] == closest_exit, 'equity'].values
-                if len(exit_equity) > 0:
-                    ax1.scatter(closest_exit, exit_equity[0], color='black', s=30, marker='o', alpha=0.6)
-        
-        ax1.set_title('Equity Curve and Performance')
-        ax1.set_ylabel('Account Value')
-        ax1.grid(True, alpha=0.3)
-        ax1.legend()
-        
-        ax2.set_title('Drawdown')
-        ax2.set_ylabel('Drawdown (%)')
-        ax2.grid(True, alpha=0.3)
-        
-        ax3.grid(True, alpha=0.3)
-        ax3.legend()
-        
-        plt.tight_layout()
-        plt.show()
-        
-        return fig
+        return plot_equity_curve_func(self)
     
     def plot_regime_performance(self):
-        if not hasattr(self, 'trade_df') or len(self.trade_df) == 0 or 'market_regime' not in self.trade_df.columns:
-            print("No regime data available.")
-            return
-        
-        import matplotlib.pyplot as plt
-        
-        regime_stats = self.trade_df.groupby('market_regime').agg({
-            'pnl': ['count', 'mean', 'sum'],
-            'position': 'count'
-        }).reset_index()
-        
-        regime_stats.columns = ['regime', 'num_trades', 'avg_pnl', 'total_pnl', 'position_count']
-        
-        regime_win_rates = []
-        for regime in regime_stats['regime'].unique():
-            regime_trades = self.trade_df[self.trade_df['market_regime'] == regime]
-            wins = sum(1 for pnl in regime_trades['pnl'] if pnl > 0)
-            total = len(regime_trades)
-            win_rate = wins / total if total > 0 else 0
-            regime_win_rates.append(win_rate * 100)
-        
-        regime_stats['win_rate'] = regime_win_rates
-        
-        fig, (ax1, ax2, ax3) = plt.subplots(1, 3, figsize=(15, 5))
-        
-        regime_colors = {
-            'strong_bull': 'green',
-            'transition_to_bull': 'lightgreen',
-            'choppy_range': 'gray',
-            'transition_to_bear': 'salmon',
-            'strong_bear': 'red',
-            'mixed': 'blue',
-            'unknown': 'lightgray'
-        }
-        
-        bars1 = ax1.bar(regime_stats['regime'], regime_stats['num_trades'], 
-                        color=[regime_colors.get(r, 'gray') for r in regime_stats['regime']])
-        ax1.set_title('Number of Trades by Market Regime')
-        ax1.set_xlabel('Market Regime')
-        ax1.set_ylabel('Number of Trades')
-        for bar in bars1:
-            height = bar.get_height()
-            ax1.annotate(f'{height}',
-                         xy=(bar.get_x() + bar.get_width() / 2, height),
-                         xytext=(0, 3),
-                         textcoords="offset points",
-                         ha='center', va='bottom')
-        
-        bars2 = ax2.bar(regime_stats['regime'], regime_stats['win_rate'], 
-                        color=[regime_colors.get(r, 'gray') for r in regime_stats['regime']])
-        ax2.set_title('Win Rate by Market Regime')
-        ax2.set_xlabel('Market Regime')
-        ax2.set_ylabel('Win Rate (%)')
-        ax2.set_ylim(0, 100)
-        for bar in bars2:
-            height = bar.get_height()
-            ax2.annotate(f'{height:.1f}%',
-                         xy=(bar.get_x() + bar.get_width() / 2, height),
-                         xytext=(0, 3),
-                         textcoords="offset points",
-                         ha='center', va='bottom')
-        
-        bars3 = ax3.bar(regime_stats['regime'], regime_stats['total_pnl'], 
-                        color=[regime_colors.get(r, 'gray') for r in regime_stats['regime']])
-        ax3.set_title('Total P&L by Market Regime')
-        ax3.set_xlabel('Market Regime')
-        ax3.set_ylabel('P&L ($)')
-        for bar in bars3:
-            height = bar.get_height()
-            ax3.annotate(f'${height:.1f}',
-                         xy=(bar.get_x() + bar.get_width() / 2, height),
-                         xytext=(0, 3),
-                         textcoords="offset points",
-                         ha='center', va='bottom')
-        
-        plt.setp(ax1.get_xticklabels(), rotation=45, ha='right')
-        plt.setp(ax2.get_xticklabels(), rotation=45, ha='right')
-        plt.setp(ax3.get_xticklabels(), rotation=45, ha='right')
-        
-        plt.tight_layout()
-        plt.show()
-        
-        return fig
+        return plot_regime_performance_func(self)
     
     def analyze_results(self):
         if self.backtest_results is None or len(self.trade_df) == 0:
